requirements.txt
streamlit
tensorflow
pillow
matplotlib
numpy
pandas
